requirements.txt
# Core Deep Learning
torch>=2.2.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
datasets>=2.18.0
accelerate>=0.28.0

# Experiment Tracking & Configuration
wandb>=0.16.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Evaluation & Metrics
seqeval>=1.2.2
scikit-learn>=1.4.0
scipy>=1.12.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Development
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Optional (faster training)
flash-attn>=2.5.0  # Optional: install manually if CUDA available
unsloth  # Optional: for S1/S2 optimization
